:sparkles: 게임 준비 로직 이관
