+++ src/main/java/baseball/Application.java
@@ -3,32 +3,31 @@ package baseball;
 import baseball.domain.BaseballNumber;
 import baseball.dto.GameStartDto;
 import baseball.enums.GameStatus;
-import baseball.service.GameReadyService;
+import baseball.service.GameInitializeService;
 import baseball.service.GameExecuteService;
-import baseball.service.impl.BaseBallGameReadyService;
+import baseball.service.impl.BaseBallGameInitializeService;
 import baseball.service.impl.BaseBallGameExecuteService;
 import baseball.service.impl.BaseBallNumberValidator;
-import baseball.util.BasicUtils;
 
 public class Application {
 
 
     public static void main(String[] args) {
-	    GameReadyService gameReadyService = new BaseBallGameReadyService(new BaseBallNumberValidator());
-	    GameExecuteService gameExecuteService = new BaseBallGameExecuteService();
+	    GameInitializeService gameInitializeService = new BaseBallGameInitializeService();
+	    GameExecuteService gameExecuteService = new BaseBallGameExecuteService(new BaseBallNumberValidator());
 	    GameStatus gameStatus;
 
 	    do{
-		    BaseballNumber systemNumber = gameReadyService.createSystemNumber();
+		    BaseballNumber systemNumber = gameInitializeService.createSystemNumber();
 			do{
 				//게임 준비
-	            GameStartDto startDto = gameReadyService.ready(systemNumber);
+	            GameStartDto startDto = gameExecuteService.ready(systemNumber);
 
 				//게임 진행
 	            gameStatus = gameExecuteService.run(startDto);
 		    } while (gameStatus.isRunning());
 
-			gameStatus = gameExecuteService.handleExecution(BasicUtils.readLine());
+			gameStatus = gameExecuteService.handleExecution();
 
 	    } while (!gameStatus.isOver());
     }
