+++ src/main/java/baseball/service/GameExecuteService.java
@@ -1,11 +1,14 @@
 package baseball.service;
 
+import baseball.domain.BaseballNumber;
 import baseball.dto.GameStartDto;
 import baseball.enums.GameStatus;
 
 public interface GameExecuteService {
 
+	GameStartDto ready(BaseballNumber systemNumber);
+
 	GameStatus run(GameStartDto gameStartDto);
 
-	GameStatus handleExecution(String inputValue);
+	GameStatus handleExecution();
 }
