+++ src/main/java/baseball/service/GameInitializeService.java
@@ -1,0 +1,8 @@
+package baseball.service;
+
+import baseball.domain.BaseballNumber;
+
+public interface GameInitializeService {
+
+	BaseballNumber createSystemNumber();
+}
