--- src/main/java/baseball/service/GameReadyService.java
@@ -1,11 +1,0 @@
-package baseball.service;
-
-import baseball.domain.BaseballNumber;
-import baseball.dto.GameStartDto;
-
-public interface GameReadyService {
-
-	GameStartDto ready(BaseballNumber systemNumber);
-
-	BaseballNumber createSystemNumber();
-}
